feat(gateway): convert rules to use actual regex (#3)
`bannedWords` is now specified as a list of regexes. They are converted to their source when constructing the final pattern (flags do not work!). Additionally, `badWordsRegex` is now compiled ahead-of-time, since I found it easy and this saves performance (and will nearly never be overeager, since it's probable that at least one non-mod-admin will send a message per session)